requirements.txt
asgiref
certifi
dj-database-url
Django
django-crispy-forms
gunicorn
pafy
pytube3
pytz
sqlparse
typing-extensions
whitenoise
youtube-dl
